wip: create base factory and unique timer device configuration objects.
